Style adminspage + refactor common components
