+++ frontend/src/components/UsersComponents/Coaches/CoachesComponents/AddCoach.tsx
@@ -2,14 +2,14 @@ import { getUsersExcludeEdition, User } from "../../../../utils/api/users/users"
 import React, { useState } from "react";
 import { addCoachToEdition } from "../../../../utils/api/users/coaches";
 import { Button, Modal, Spinner } from "react-bootstrap";
-import { Error } from "../../Requests/styles";
-import { AddCoachButtonDiv, ModalContentConfirm } from "../../../AdminsComponents/styles";
+import { AddButtonDiv } from "../../../AdminsComponents/styles";
 import { AsyncTypeahead, Menu } from "react-bootstrap-typeahead";
 import UserMenuItem from "../../../Common/Users/MenuItem";
-import { StyledMenuItem } from "../../../Common/Users/styles";
+import { Error, StyledMenuItem } from "../../../Common/Users/styles";
 import { EmailAndAuth } from "../../../Common/Users";
 import { EmailDiv } from "../styles";
 import CreateButton from "../../../Common/Buttons/CreateButton";
+import { ModalContentConfirm } from "../../../Common/styles";
 
 /**
  * A button and popup to add a new coach to the given edition.
@@ -102,11 +102,11 @@ export default function AddCoach(props: { edition: string; refreshCoaches: () =>
 
     return (
         <>
-            <AddCoachButtonDiv>
+            <AddButtonDiv>
                 <CreateButton showIcon={false} onClick={handleShow}>
                     Add coach
                 </CreateButton>
-            </AddCoachButtonDiv>
+            </AddButtonDiv>
 
             <Modal show={show} onHide={handleClose}>
                 <ModalContentConfirm>
