+++ frontend/src/components/UsersComponents/Coaches/styles.ts
@@ -36,14 +36,6 @@ export const RemoveTd = styled.td`
     vertical-align: middle;
 `;
 
-export const ListDiv = styled.div`
-    width: 100%;
-    height: fit-content;
-    max-height: 40em;
-    overflow: auto;
-    margin-top: 10px;
-`;
-
 export const DialogButtonDiv = styled.div`
     margin-right: 4px;
     margin-bottom: 4px;
@@ -64,14 +56,4 @@ export const EmailDiv = styled.div`
 export const CredsDiv = styled.div`
     overflow: hidden;
     text-overflow: ellipsis;
-`;
-
-export const SearchFieldDiv = styled.div`
-    float: left;
-    width: 15em;
-`;
-
-export const TableDiv = styled.div`
-    width: 100%;
-    clear: left;
 `;
