+++ frontend/src/views/AdminsPage/styles.ts
@@ -5,7 +5,3 @@ export const AdminsContainer = styled.div`
     max-width: 90%;
     margin: 10px auto auto;
 `;
-
-export const AdminsHeader = styled.div`
-    margin-bottom: 5px;
-`;
